refactor: Move interactive packages under runtime namespace
