+++ cmd/app/addDiscord.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/bwmarrin/discordgo"
 	"github.com/manifoldco/promptui"
-	"github.com/pardnchiu/agenvoy/internal/interactive/discord"
+	"github.com/pardnchiu/agenvoy/internal/runtime/discord"
 	"github.com/pardnchiu/agenvoy/internal/session"
 	"github.com/pardnchiu/go-pkg/filesystem/keychain"
 	"golang.org/x/term"
