+++ internal/runtime/discord/command/handler.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"log/slog"
 
-	discordTypes "github.com/pardnchiu/agenvoy/internal/interactive/discord/types"
+	discordTypes "github.com/pardnchiu/agenvoy/internal/runtime/discord/types"
 )
 
 func Handler(receiveMessage *discordTypes.ReceiveMessage) []discordTypes.ReplyMessage {
