+++ internal/runtime/discord/new.go
@@ -10,8 +10,8 @@ import (
 	"github.com/bwmarrin/discordgo"
 	"github.com/pardnchiu/agenvoy/internal/agents/host"
 	agentTypes "github.com/pardnchiu/agenvoy/internal/agents/types"
-	discordCommand "github.com/pardnchiu/agenvoy/internal/interactive/discord/command"
-	discordTypes "github.com/pardnchiu/agenvoy/internal/interactive/discord/types"
+	discordCommand "github.com/pardnchiu/agenvoy/internal/runtime/discord/command"
+	discordTypes "github.com/pardnchiu/agenvoy/internal/runtime/discord/types"
 	"github.com/pardnchiu/agenvoy/internal/scheduler"
 	"github.com/pardnchiu/agenvoy/internal/session"
 	"github.com/pardnchiu/go-pkg/filesystem/keychain"
