+++ internal/runtime/tui/commandMcp.go
