+++ internal/runtime/tui/commandMcpAdd.go
