+++ internal/runtime/tui/commandMcpRemove.go
