+++ internal/runtime/tui/commandMode.go
