+++ internal/runtime/tui/commandModel.go
