+++ internal/runtime/tui/commandModelAdd.go
