+++ internal/runtime/tui/commandPlanner.go
