+++ internal/runtime/tui/commandReasoning.go
