+++ internal/runtime/tui/commandSessionModel.go
