+++ internal/runtime/tui/commandSwitch.go
