+++ internal/runtime/tui/commnadModelRemove.go
