+++ internal/runtime/tui/handlerCommand.go
