+++ internal/runtime/tui/handlerLogFormat.go
