+++ internal/runtime/tui/handlerLogMode.go
