+++ internal/runtime/tui/view.go
