+++ internal/runtime/tui/viewRender.go
